Rename module and update mirrors (gh issue #98).
